+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * `SelectList` inherits parent width more naturally, and no longer requires an explicit numeric width be set
 * `SelectList` supports `inverted` prop, to expand the selection options upwards rather than downwards
 * `Tag` enables transitions, for consistency with other components
+* Standardize transitions across all components with custom easing function
 
 ## v1.2.1